Rebuild settings tree when changing the ScriptPath
This patch fixes bugs related to changing the script path and having to close and reopen the window to see the updated settings tree.
This is definitely not a very good way of doing this as an autosplitter could register their settings not on the first update loop but I suppose it works for now.