Implement variable refresh rate for autosplitters
This is built off the tick rate helper added in the last commit, when there's a change in the tick rate, the component will detect it and change it's own refresh timer.
This means that calling set_tick_rate in a autosplitter will now have an effect.
This timer approach is very similar to the ASL component, I think it is improvable but it works.