Fix various autosplitter reloading issues
The variable refresh rate refactor introduced some problems with reloading scripts and UI not updating correctly, this fixes them.

We are now handling the script reloading and the execution in 2 different threads:
- The script reloading happens in the UI thread to allow for proper UI refresh
- The script execution happens in the external timer

I am not 100% sure if there can be a problem with race conditions but it seems like it works fine for me.

While this isn't very elegant, this is now try finally the whole update process, whatever happens in the script execution, we want to reenable the timer which is the main reason why we're doing this.