Completely refactor the settings and add choices
This completely refactors the settings, ensuring they more cleanly match
the way the ASR API works. We don't have a dictionary for the settings
anymore and instead directly hold the settings map. Additionally this
completely remove the checkbox tree widget and now models an actual GUI
that is being built. With the updated `livesplit-core` it's also now
possible to have i64, f64, string, and list settings. Also there is now
a choice widget that is shown as a combo box.